fastapi-service/pyproject.toml
[project]
name = "fastapi-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "pypdf>=6.10.0",
    "python-multipart>=0.0.26",
    "sentence-transformers>=5.4.0",
    "uvicorn>=0.44.0",
]
